refactor : 카카오 리다이렉트 페이지 컴포넌트화
